+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this
 ## [1.7.1] - Unreleased
 
 ### Fixed
+- Prevented clipboard contention from crashing MAC address, IP address, and hostname copying.
 - Selected device rows now use their theme's established selection colors.
 
 ## [1.7.0] - 2026-09-05